* v1.37g, 12 May 2006       + Added Font menu to TextWindows ("Results", "Log", etc.).       + The Edit/Options/Font window (or double click on text tool) now         lists all available fonts.       + On Mac OS X, opening images using drag and drop is faster.       + Added the Macros/Evaluate Line command to the editor.       + In the "Log" window, double click on paths to open Java and macro         source files (Use with Find and Search_Macros macros).       + Single lines copied from TextWindows ("Results", "Log", etc.) no         longer have \n at the end.       + The maximum size of the macro symbol table increased from 32K to         512K.       + Fixed a bug that could cause calculated centroid coordinates to be         incorrect with large images.       + Fixed a bug in the image calculator that caused ANDing of an RGB         image and an 8-bit mask to not work correctly if the mask had an         inverting LUT.

+++ ij/process/FloatStatistics.java
@@ -186,8 +186,8 @@ public class FloatStatistics extends ImageStatistics {
 				i++;
 			}
 		}
-		xCentroid = (double)xsum/count+0.5;
-		yCentroid = (double)ysum/count+0.5;
+		xCentroid = xsum/count+0.5;
+		yCentroid = ysum/count+0.5;
 		if (cal!=null) {
 			xCentroid = cal.getX(xCentroid);
 			yCentroid = cal.getY(yCentroid, height);
